+++ examples/flowcharts/VerticalFlowChart.tsx
@@ -75,7 +75,7 @@ export const VerticalFlowChart = ({ fref, chartData }: MilestoneStory) => {
                         1. Beginning
                     </BoxedLabel>
                 </MilestoneMotion>
-                <MilestoneMotion enterOn={'A'} transition={{ delay: 0.5 }} exit={null}>
+                <MilestoneMotion enterOn={'A'} enterTransition={{ delay: 0.5 }}>
                     <Label position={[145, y1]} size={[200, 20]} className={'explanation'}>
                         Usually, we start at the beginning.
                     </Label>
@@ -90,17 +90,17 @@ export const VerticalFlowChart = ({ fref, chartData }: MilestoneStory) => {
                         markerEnd={'arrowV'}
                     />
                 </MilestoneMotion>
-                <MilestoneMotion enterOn={'B'} transition={{ delay: 0.5 }} exit={null}>
+                <MilestoneMotion enterOn={'B'} enterTransition={{ delay: 0.5 }}>
                     <BoxedLabel position={[60, y2]} size={[120, 28]} className={'B'}>
                         2. Middle
                     </BoxedLabel>
                 </MilestoneMotion>
-                <MilestoneMotion enterOn={'B'} transition={{ delay: 1 }} exit={null}>
+                <MilestoneMotion enterOn={'B'} enterTransition={{ delay: 1 }}>
                     <Label position={[145, y2 - 10]} size={[200, 20]} className={'explanation'}>
                         The middle stage is often the longest.
                     </Label>
                 </MilestoneMotion>
-                <MilestoneMotion enterOn={'B'} transition={{ delay: 1.5 }} exit={null}>
+                <MilestoneMotion enterOn={'B'} enterTransition={{ delay: 1.5 }}>
                     <Label position={[145, y2 + 10]} size={[200, 20]} className={'explanation'}>
                         It may contain many sub-stages.
                     </Label>
@@ -115,12 +115,12 @@ export const VerticalFlowChart = ({ fref, chartData }: MilestoneStory) => {
                         markerEnd={'arrowV'}
                     />
                 </MilestoneMotion>
-                <MilestoneMotion enterOn={'C'} transition={{ delay: 0.5 }} exit={null}>
+                <MilestoneMotion enterOn={'C'} enterTransition={{ delay: 0.5 }}>
                     <BoxedLabel position={[60, y3]} size={[120, 28]} className={'C'}>
                         3. End
                     </BoxedLabel>
                 </MilestoneMotion>
-                <MilestoneMotion enterOn={'C'} transition={{ delay: 1 }} exit={null}>
+                <MilestoneMotion enterOn={'C'} enterTransition={{ delay: 1 }}>
                     <Label position={[145, y3]} size={[200, 20]} className={'explanation'}>
                         Finally, we reach a conclusion.
                     </Label>