+++ packages/core/src/themes/hooks.tsx
@@ -1,6 +1,7 @@
 import { useTheme } from './context'
 import { fillProps } from './utils'
-import { ThemedComponent, WithVariant } from './types'
+import { ThemedComponent } from './types'
+import { WithVariant } from '../general/types'
 
 export const useThemedProps = <Props extends WithVariant>(props: Props, key: ThemedComponent) => {
     const theme = useTheme()