+++ packages/core/src/typography/Counter.tsx
@@ -1,22 +1,21 @@
 import { animate } from 'framer-motion'
 import { createElement, useState } from 'react'
 import { roundDecimalPlaces } from '../general'
-import { TransitionProps, useTheme } from '../themes'
 import { CounterProps } from './types'
 import { Label } from './Label'
 
 const CounterContent = ({
-    config,
+    transition,
     nDecimalPlaces = 0,
     format = (v: number) => String(v),
     children,
-}: Pick<CounterProps, 'format' | 'nDecimalPlaces' | 'children'> & { config: TransitionProps }) => {
+}: Pick<CounterProps, 'transition' | 'nDecimalPlaces' | 'format' | 'children'>) => {
     const [value, setValue] = useState(roundDecimalPlaces(Number(children), nDecimalPlaces))
     const [working, setWorking] = useState(false)
 
     if (value !== Number(children) && !working) {
         animate(value, Number(children), {
-            ...config,
+            ...transition,
             onPlay: () => {
                 setWorking(true)
             },
@@ -37,13 +36,12 @@ export const Counter = ({
     nDecimalPlaces,
     format,
     component = Label,
+    transition,
     children,
     ...props
 }: CounterProps) => {
-    const theme = useTheme()
-    const config = theme.Motion[variant] ?? theme.Motion.default
     const content = (
-        <CounterContent config={config} nDecimalPlaces={nDecimalPlaces} format={format}>
+        <CounterContent transition={transition} nDecimalPlaces={nDecimalPlaces} format={format}>
             {children}
         </CounterContent>
     )